[feature] 그림 그리기 폼 기능 구현
