+++ src/features/drawing/types/drawing.ts
@@ -1,0 +1,7 @@
+export type Tool = 'pen' | 'eraser' | 'fill';
+
+export interface Lines {
+  tool: Tool;
+  points: number[];
+  color: string;
+}
